% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_keep_label}
\alias{filter_keep_label}
\title{Filter a dataset, keeping attributes - fix me what if attr assignment fails}
\usage{
filter_keep_label(df, split, addlabel = " - subset")
}
\arguments{
\item{df}{A data frame you want to split}

\item{split}{a vector which will filter the data frame. NA values replaced as FALSE}

\item{addlabel}{A note you should add to the label attribute so that the filtering is explained in graphs etc}
}
\value{
The filtered data frame, usually with fewer rows. But the attributes of the variables are retained where possible.
}
\description{
Filter a dataset, keeping attributes - fix me what if attr assignment fails
}

