% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{ggplotl}
\alias{ggplotl}
\title{Simple wrapper for ggplot
So you can print the labels instead of the varnames
Set a label:
\code{attr(mtcars$cyl,"label")="cylinder"}
Or, because reprojuicer imports Hmisc, more conveniently:
\code{label(mtcars$cyl)="cylinder"}
\code{library(ggplot2);ggplotl(data=mtcars,aes(cyl,disp))+geom_point()}}
\usage{
ggplotl(..., labs = list())
}
\arguments{
\item{labs}{If you want to use attributes other than \code{label} for your labels, give it a list like \code{labs=list(x="otherattribute",y="yetanotherattribute"))}. By default, the attribute \code{label} is used.}
}
\value{
various things
}
\description{
Simple wrapper for ggplot
So you can print the labels instead of the varnames
Set a label:
\code{attr(mtcars$cyl,"label")="cylinder"}
Or, because reprojuicer imports Hmisc, more conveniently:
\code{label(mtcars$cyl)="cylinder"}
\code{library(ggplot2);ggplotl(data=mtcars,aes(cyl,disp))+geom_point()}
}
\examples{
attr(mtcars$disp,"labelb")="čćđž"
attr(mtcars$disp,"label")="Displacement"
library(reprojuicer)
ggplotl(data=mtcars,aes(disp))+geom_density()
ggplotl(data=mtcars,aes(disp),labs=list(x="labelb"))+geom_density()
ggplotl(data=mtcars,aes(gear,cyl,colour=disp),labs=list(colour="labelb"))+geom_point()
}

