% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{reprojuice}
\alias{reprojuice}
\title{Provides an easy xls workbook for reproducibly organising a dataset - reordering, renaming, recoding, providing longer and shorter labels, switching easily between labels in multiple languages, calculating new variables, grouping into blocks.}
\usage{
reprojuice(dataname = "data", type = "", method = "quicker",
  onlyLoad = F, spreadsheet_path = "")
}
\description{
Main intro at the moment is the readme: \link{https://github.com/stevepowell99/reprojuicer}
Note you can use raw_data$myvar in formula to refer to version of an earlier variable which has already been changed by a formula.
Use label2 for another language, create more if you want
}

