% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{ggplotl}
\alias{ggplotl}
\title{Simple wrapper for ggplot
So you can print the labels instead of the varnames}
\usage{
ggplotl(...)
}
\arguments{
\item{...}{same as ggplot}
}
\value{
various things


Set a label:
\code{attr(mtcars$cyl,"label")="cylinder"}
Or, because reprojuicer imports Hmisc, more conveniently:
\code{label(mtcars$cyl)="cylinder"}
\code{library(ggplot2);ggplotl(data=mtcars,aes(cyl,disp))+geom_point()}
}
\description{
Simple wrapper for ggplot
So you can print the labels instead of the varnames
}

